# docs/swagger.yaml
basePath: /
definitions:
...
  godevmandb.CreateDeviceTypeParams:
    properties:
      class_id:
        type: integer
      hc:
        type: boolean
      manufacturer:
        type: string
      model:
        type: string
      snmp_ver:
        type: integer
      sys_id:
        type: string
    type: object
...
  godevmandb.DeviceClass:
    properties:
      class_id:
        type: integer
      created_on:
        type: string
      descr:
        type: string
      updated_on:
        type: string
    type: object
...
  godevmandb.DeviceType:
    properties:
      class_id:
        type: integer
      created_on:
        type: string
      hc:
        type: boolean
      manufacturer:
        type: string
      model:
        type: string
      snmp_ver:
        description: highest supported snmp version
        type: integer
      sys_id:
        description: snmp sysObjectId or some unique identifier if snmp is not supported
        type: string
      updated_on:
        type: string
    type: object
...
  godevmandb.UpdateDeviceTypeParams:
    properties:
      class_id:
        type: integer
      hc:
        type: boolean
      manufacturer:
        type: string
      model:
        type: string
      snmp_ver:
        type: integer
      sys_id:
        type: string
    type: object
  handlers.CountResponse:
    properties:
      count:
        type: integer
    type: object
  handlers.StatusResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  handlers.device:
    properties:
      backup:
        type: boolean
      backup_failed:
        type: boolean
      created_on:
        type: string
      dev_id:
        type: integer
      dom_id:
        type: integer
      ext_model:
        type: string
      graph:
        type: boolean
      host_name:
        type: string
      installed:
        type: boolean
      ip4_addr:
        type: string
      ip6_addr:
        type: string
      monitor:
        type: boolean
      notes:
        type: string
      parent:
        type: integer
      site_id:
        type: integer
      snmp_main_id:
        type: integer
      snmp_ro_id:
        type: integer
      source:
        type: string
      sw_version:
        type: string
      sys_contact:
        type: string
      sys_id:
        type: string
      sys_location:
        type: string
      sys_name:
        type: string
      type_changed:
        type: boolean
      unresponsive:
        type: boolean
      updated_on:
        type: string
      validation_failed:
        type: boolean
    type: object
...
info:
  contact:
    email: marko@aretaja.org
    name: API Support
  description: goDevmans API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: goDevmanAPI
  version: v0.0.1-devel.1
paths:
...
  /devices/{dev_id}:
    delete:
      description: Delete device
      operationId: delete-device
      parameters:
      - description: dev_id
        in: path
        name: dev_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid dev_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Delete device
      tags:
      - devices
    get:
      description: Get device info
      operationId: get-device
      parameters:
      - description: dev_id
        in: path
        name: dev_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.device'
        "400":
          description: Invalid dev_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Get device
      tags:
      - devices
    put:
      description: Update device
      operationId: update-device
      parameters:
      - description: dev_id
        in: path
        name: dev_id
        required: true
        type: string
      - description: JSON object of device.<br />Ignored fields:<ul><li>dev_id</li><li>updated_on</li><li>created_on</li></ul>
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/handlers.device'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.device'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Update device
      tags:
      - devices
...
  /devices/types:
    get:
      description: List device types info
      operationId: list-device_types
      parameters:
      - description: url encoded SQL 'ILIKE' operator pattern
        in: query
        name: sys_id_f
        type: string
      - description: url encoded SQL 'ILIKE' operator pattern
        in: query
        name: manufacturer_f
        type: string
      - description: url encoded SQL 'ILIKE' operator pattern
        in: query
        name: model_f
        type: string
      - description: 'min: 1; max: 1000; default: 100'
        in: query
        name: limit
        type: integer
      - description: 'default: 0'
        in: query
        name: offset
        type: integer
      - description: record update time >= (unix timestamp in milliseconds)
        in: query
        name: updated_ge
        type: integer
      - description: record update time <= (unix timestamp in milliseconds)
        in: query
        name: updated_le
        type: integer
      - description: record creation time >= (unix timestamp in milliseconds)
        in: query
        name: created_ge
        type: integer
      - description: record creation time <= (unix timestamp in milliseconds)
        in: query
        name: created_le
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/godevmandb.DeviceType'
            type: array
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: List device_types
      tags:
      - devices
    post:
      description: Create device type
      operationId: create-device_type
      parameters:
      - description: JSON object of godevmandb.CreateDeviceTypeParams<br />sys_id
          must match ^[\w-\.]+$ regex
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/godevmandb.CreateDeviceTypeParams'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/godevmandb.DeviceType'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Create device_type
      tags:
      - devices
  /devices/types/{sys_id}:
    delete:
      description: Delete device type
      operationId: delete-device_type
      parameters:
      - description: sys_id
        in: path
        name: sys_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid sys_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Delete device_type
      tags:
      - devices
    get:
      description: Get device type info
      operationId: get-device_type
      parameters:
      - description: sys_id
        in: path
        name: sys_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/godevmandb.DeviceType'
        "400":
          description: Invalid sys_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Domain not found
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Get device_type
      tags:
      - devices
    put:
      description: Update device type
      operationId: update-device_type
      parameters:
      - description: sys_id
        in: path
        name: sys_id
        required: true
        type: string
      - description: JSON object of godevmandb.UpdateDeviceTypeParams.<br />Ignored
          fields:<ul><li>sys_id</li></ul>
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/godevmandb.UpdateDeviceTypeParams'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/godevmandb.DeviceType'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Update device_type
      tags:
      - devices
  /devices/types/{sys_id}/class:
    get:
      description: Get device type class info
      operationId: get-device-type-class
      parameters:
      - description: sys_id
        in: path
        name: sys_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/godevmandb.DeviceClass'
        "400":
          description: Invalid sys_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Get device type class
      tags:
      - devices
  /devices/types/count:
    get:
      description: Count number of device types
      operationId: count-device_types
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CountResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Count device_types
      tags:
      - devices
...
swagger: "2.0"